frontend: resourceMap: Expose GraphView and KubeIcon through pluginLib
Relates to issue 6556. Plugins had no way to embed the Map's graph
renderer or icon set on their own detail pages, since resourceMap/
was never re-exported through pluginLib. Adding GraphView to the
export list directly would have broken though, because
KubeObjectNode.tsx and GroupNode.tsx imported useGraphView/useNode
back from GraphView.tsx, creating GraphView -> GraphRenderer ->
nodes/*.tsx -> GraphView circular import. That ordering happens to
resolve inside a single bundle, but the guarantee doesn't carry over
once GraphView is re-exported through a separate pluginLib entry
point, leaving the hooks undefined at import time for plugin bundles.

Move GraphViewContext, FullGraphContext, useGraphView, useNode into
a new graphViewContext.tsx with no dependency on GraphView.tsx or the
node renderers, and have both sides import from there instead of each
other. GraphView.tsx re-exports them for existing consumers.

Add a resourceMap/index.ts barrel exporting GraphView, KubeIcon,
getKindGroupColor, the graph context/hooks and the GraphNode/GraphEdge/
GraphSource/Relation/GraphFilter types, and attach it to
window.pluginLib.ResourceMap in plugin/index.ts, following the same
pattern already used for CommonComponents.

Plugins built with @kinvolk/headlamp-plugin can now import from
'@kinvolk/headlamp-plugin/lib/ResourceMap' (or
'.../lib/components/resourceMap'), which resolves to
pluginLib.ResourceMap via the externalModules map in
plugins/headlamp-plugin/config/vite.config.mjs.
