+++ frontend/src/components/resourceMap/GraphView.tsx
@@ -24,11 +24,9 @@ import { styled } from '@mui/material/styles';
 import ThemeProvider from '@mui/system/ThemeProvider';
 import { Edge, Node, Panel, ReactFlowProvider } from '@xyflow/react';
 import {
-  createContext,
   ReactNode,
   StrictMode,
   useCallback,
-  useContext,
   useEffect,
   useLayoutEffect,
   useMemo,
@@ -53,7 +51,7 @@ import {
 } from './graph/graphGrouping';
 import { detectGraphChanges, shouldUseIncrementalUpdate } from './graph/graphIncrementalUpdate';
 import { applyGraphLayout } from './graph/graphLayout';
-import { GraphLookup, makeGraphLookup } from './graph/graphLookup';
+import { makeGraphLookup } from './graph/graphLookup';
 import { forEachNode, GraphEdge, GraphNode, GraphSource, Relation } from './graph/graphModel';
 import {
   EXTREME_SIMPLIFICATION_THRESHOLD,
@@ -64,6 +62,7 @@ import {
 } from './graph/graphSimplification';
 import { GraphControlButton } from './GraphControls';
 import { GraphRenderer } from './GraphRenderer';
+import { FullGraphContext, GraphViewContext } from './graphViewContext';
 import { PerformanceStats } from './PerformanceStats';
 import { SelectionBreadcrumbs } from './SelectionBreadcrumbs';
 import { useGetAllRelations } from './sources/definitions/relations';
@@ -73,25 +72,16 @@ import { GraphSourcesView } from './sources/GraphSourcesView';
 import { useGraphViewport } from './useGraphViewport';
 import { useQueryParamsState } from './useQueryParamsState';
 
-interface GraphViewContent {
-  setNodeSelection: (nodeId: string) => void;
-  nodeSelection?: string;
-}
-export const GraphViewContext = createContext({} as any);
-export const useGraphView = () => useContext<GraphViewContent>(GraphViewContext);
-
-interface FullGraphContent {
-  fullGraph: any;
-  lookup: GraphLookup<GraphNode, GraphEdge>;
-}
-export const FullGraphContext = createContext({} as any);
-export const useFullGraphContext = () => useContext<FullGraphContent>(FullGraphContext);
-
-export const useNode = (id: string) => {
-  const { lookup } = useFullGraphContext();
-
-  return lookup.getNode(id);
-};
+// Re-exported here for backwards compatibility with existing consumers that
+// imported these from GraphView.tsx. The actual definitions live in
+// graphViewContext.tsx to avoid a circular import with the node renderers.
+export {
+  FullGraphContext,
+  GraphViewContext,
+  useFullGraphContext,
+  useNode,
+} from './graphViewContext';
+export { useGraphView } from './graphViewContext';
 
 interface GraphViewContentProps {
   /** Height of the Map */
