+++ plugins/headlamp-plugin/src/index.ts
@@ -22,6 +22,7 @@ declare module '@mui/private-theming' {
 
 import { Activity } from './components/activity/Activity';
 import * as CommonComponents from './components/common';
+import * as ResourceMap from './components/resourceMap';
 import type { AppTheme } from './lib/AppTheme';
 import * as K8s from './lib/k8s';
 import * as ApiProxy from './lib/k8s/apiProxy';
@@ -88,6 +89,7 @@ export {
   K8s,
   K8s as k8s,
   CommonComponents,
+  ResourceMap,
   Utils,
   Router,
   Plugin,
